AMP-7515
git-svn-id: file:///svn/branches/AMP_2_0_RELEASE@13843 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/util/LuceneUtil.java
@@ -234,17 +234,18 @@ public class LuceneUtil implements Serializable {
     		}
     	}
     	
-    	if (idxDir.exists()){
-			logger.info("Found, deleting ...");
-			if (deleteDirectory(idxDir))
-				logger.info("Done");
-			else
-				logger.info("Can't delete!");
-		}
-		else{
-			logger.info("Not found ... will be generated!");
-		}
-
+    	if (deleteIndex){
+    		if (idxDir.exists()){
+    			logger.info("Found, deleting ...");
+    			if (deleteDirectory(idxDir))
+    				logger.info("Done");
+    			else
+    				logger.info("Can't delete!");
+    		}
+    		else{
+    			logger.info("Not found ... will be generated!");
+    		}
+    	}
    	
     	if (!idxDir.exists()){ //we need to create the index from 0  
     		if (!idxDir.mkdirs()){
